Updates to structure and data files to support peaking calculations. Peaks now calculated based on ratio of peak seasonal demand by end use to average annual peak demand (we found a new data source to allow for this estimation). We no longer need the seasonal share of energy consutmpion so I've removed the folder structure and data files.
The results in early years are reasonable and I'm working on updating the data ensure the later years produce reasonable results (right now our peak demand is a bit too high).

I also reinstated RM because the peak demand numbers came more in line with actual demand, meaning the RM is back to serving a purpose (and not incorporated into the underying data).